requirements.txt
alembic==1.4.2
asn1crypto==1.3.0
Babel==2.8.0
blinker==1.4
cffi==1.14.0
click==7.1.1
#cryptography==2.9
dominate==2.5.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.2.1